requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.1
billiard==3.3.0.23
celery==3.1.18
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
idna==2.10
kombu==3.0.37
mysqlclient==2.0.3
phonenumbers==8.12.19
PyJWT==1.7.1
pytz==2021.1
redis==2.10.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
twilio==6.53.0
urllib3==1.26.3
